Update Aetheris branding, move Terminal to side menu, and update notifications with sound

+++ app/src/main/java/com/breadwallet/tools/manager/BRSharedPrefs.java
@@ -38,7 +38,7 @@ public class BRSharedPrefs {
     }
 
     public static String getIsoSymbol(Context context) {
-
+        if (context == null) return "USD";
         SharedPreferences settingsToGet = context.getSharedPreferences(BRConstants.PREFS_NAME, 0);
         String defIso;
         String defaultLanguage = Locale.getDefault().getLanguage();
@@ -61,6 +61,7 @@ public class BRSharedPrefs {
     }
 
     public static void putIso(Context context, String code) {
+        if (context == null) return;
         SharedPreferences settings = context.getSharedPreferences(BRConstants.PREFS_NAME, 0);
         SharedPreferences.Editor editor = settings.edit();
         editor.putString(BRConstants.CURRENT_CURRENCY, code.equalsIgnoreCase(Locale.getDefault().getISO3Language()) ? null : code);
@@ -78,43 +79,51 @@ public class BRSharedPrefs {
     //////////////////////////////////////////////////////////////////////////////
     //////////////////// Active Shared Preferences ///////////////////////////////
     public static void putLastSyncTimestamp(Context activity, long time) {
+        if (activity == null) return;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putLong("lastSyncTime", time);
         editor.apply();
     }
     public static long getLastSyncTimestamp(Context activity) {
+        if (activity == null) return 0;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getLong("lastSyncTime", 0L);
     }
     public static void putStartSyncTimestamp(Context activity, long time) {
+        if (activity == null) return;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putLong("startSyncTime", time);
         editor.apply();
     }
     public static long getStartSyncTimestamp(Context activity) {
+        if (activity == null) return 0;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getLong("startSyncTime", 0L);
     }
 
     public static void putSyncTimeElapsed(Context activity, long time) {
+        if (activity == null) return;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putLong("syncTimeElapsed", time);
         editor.apply();
     }
     public static long getSyncTimeElapsed(Context activity) {
+        if (activity == null) return 0;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getLong("syncTimeElapsed", 0L);
     }
 
     public static boolean getPhraseWroteDown(Context context) {
+        if (context == null) return false;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getBoolean(BRConstants.PHRASE_WRITTEN, false);
     }
 
     public static void putPhraseWroteDown(Context context, boolean check) {
+        if (context == null) return;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putBoolean(BRConstants.PHRASE_WRITTEN, check);
@@ -122,11 +131,13 @@ public class BRSharedPrefs {
     }
 
     public static boolean getGreetingsShown(Context context) {
+        if (context == null) return false;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getBoolean("greetingsShown", false);
     }
 
     public static void putGreetingsShown(Context context, boolean shown) {
+        if (context == null) return;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putBoolean("greetingsShown", shown);
@@ -134,11 +145,13 @@ public class BRSharedPrefs {
     }
 
     public static int getCurrencyListPosition(Context context) {
+        if (context == null) return 0;
         SharedPreferences settings = context.getSharedPreferences(BRConstants.PREFS_NAME, 0);
         return settings.getInt(BRConstants.POSITION, 0);
     }
 
     public static void putCurrencyListPosition(Context context, int lastItemsPosition) {
+        if (context == null) return;
         SharedPreferences settings = context.getSharedPreferences(BRConstants.PREFS_NAME, 0);
         SharedPreferences.Editor editor = settings.edit();
         editor.putInt(BRConstants.POSITION, lastItemsPosition);
@@ -146,22 +159,26 @@ public class BRSharedPrefs {
     }
 
     public static String getReceiveAddress(Context context) {
+        if (context == null) return "";
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getString(BRConstants.RECEIVE_ADDRESS, "");
     }
 
     public static void putReceiveAddress(Context ctx, String tmpAddr) {
+        if (ctx == null) return;
         SharedPreferences.Editor editor = ctx.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE).edit();
         editor.putString(BRConstants.RECEIVE_ADDRESS, tmpAddr);
         editor.apply();
     }
 
     public static String getFirstAddress(Context context) {
+        if (context == null) return "";
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getString(BRConstants.FIRST_ADDRESS, "");
     }
 
     public static void putFirstAddress(Context context, String firstAddress) {
+        if (context == null) return;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putString(BRConstants.FIRST_ADDRESS, firstAddress);
@@ -169,11 +186,13 @@ public class BRSharedPrefs {
     }
 
     public static long getCatchedBalance(Context context) {
+        if (context == null) return 0;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getLong("balance", 0);
     }
 
     public static void putCatchedBalance(Context context, long fee) {
+        if (context == null) return;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putLong("balance", fee);
@@ -181,12 +200,14 @@ public class BRSharedPrefs {
     }
 
     public static long getSecureTime(Context activity) {
+        if (activity == null) return System.currentTimeMillis() / 1000;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getLong(BRConstants.SECURE_TIME_PREFS, System.currentTimeMillis() / 1000);
     }
 
     //secure time from the server
     public static void putSecureTime(Context activity, long date) {
+        if (activity == null) return;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putLong(BRConstants.SECURE_TIME_PREFS, date);
@@ -194,11 +215,13 @@ public class BRSharedPrefs {
     }
 
     public static long getFeeTime(Context activity) {
+        if (activity == null) return 0;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getLong("feeTime", 0);
     }
 
     public static void putFeeTime(Context activity, long feeTime) {
+        if (activity == null) return;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putLong("feeTime", feeTime);
@@ -206,11 +229,13 @@ public class BRSharedPrefs {
     }
 
     public static boolean getAllowSpend(Context activity) {
+        if (activity == null) return true;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getBoolean(BRConstants.ALLOW_SPEND, true);
     }
 
     public static void putAllowSpend(Context activity, boolean allow) {
+        if (activity == null) return;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putBoolean(BRConstants.ALLOW_SPEND, allow);
@@ -219,12 +244,14 @@ public class BRSharedPrefs {
 
     //if the user prefers all in litecoin units, not other currencies
     public static boolean getPreferredLTC(Context activity) {
+        if (activity == null) return true;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getBoolean("priceSetToLitecoin", true);
     }
 
     //if the user prefers all in litecoin units, not other currencies
     public static void putPreferredLTC(Context activity, boolean b) {
+        if (activity == null) return;
         Timber.d("timber: putPreferredLTC: %s", b);
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
@@ -234,18 +261,21 @@ public class BRSharedPrefs {
 
     //if the user prefers all in litecoin units, not other currencies
     public static boolean getUseFingerprint(Context activity) {
+        if (activity == null) return false;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getBoolean("useFingerprint", false);
     }
 
     //if the user prefers all in litecoin units, not other currencies
     public static void putUseFingerprint(Context activity, boolean use) {
+        if (activity == null) return;
         SharedPreferences prefs = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putBoolean("useFingerprint", use);
         editor.apply();
     }
     public static int getStartHeight(Context context) {
+        if (context == null) return 0;
         SharedPreferences settingsToGet = context.getSharedPreferences(BRConstants.PREFS_NAME, 0);
         return  settingsToGet.getInt(BRConstants.START_HEIGHT, 0);
     }
@@ -259,6 +289,7 @@ public class BRSharedPrefs {
     }
 
     public static int getLastBlockHeight(Context context) {
+        if (context == null) return 0;
         SharedPreferences settingsToGet = context.getSharedPreferences(BRConstants.PREFS_NAME, 0);
         return settingsToGet.getInt(BRConstants.LAST_BLOCK_HEIGHT, 0);
     }
@@ -272,6 +303,7 @@ public class BRSharedPrefs {
     }
 
     public static boolean getScanRecommended(Context context) {
+        if (context == null) return false;
         SharedPreferences settingsToGet = context.getSharedPreferences(BRConstants.PREFS_NAME, 0);
         return settingsToGet.getBoolean("scanRecommended", false);
     }
@@ -285,6 +317,7 @@ public class BRSharedPrefs {
     }
 
     public static int getCurrencyUnit(Context context) {
+        if (context == null) return BRConstants.CURRENT_UNIT_LITECOINS;
         SharedPreferences settingsToGet = context.getSharedPreferences(BRConstants.PREFS_NAME, 0);
         return settingsToGet.getInt(BRConstants.CURRENT_UNIT, BRConstants.CURRENT_UNIT_LITECOINS);
     }
@@ -306,12 +339,14 @@ public class BRSharedPrefs {
     }
 
     public static void clearAllPrefs(Context activity) {
+        if (activity == null) return;
         SharedPreferences.Editor editor = activity.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE).edit();
         editor.clear();
         editor.apply();
     }
 
     public static boolean getShowNotification(Context context) {
+        if (context == null) return false;
         SharedPreferences settingsToGet = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return settingsToGet.getBoolean("showNotification", false);
     }
@@ -325,6 +360,7 @@ public class BRSharedPrefs {
     }
 
     public static boolean getShareData(Context context) {
+        if (context == null) return false;
         SharedPreferences settingsToGet = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return settingsToGet.getBoolean("shareData", false);
     }
@@ -338,6 +374,7 @@ public class BRSharedPrefs {
     }
 
     public static boolean getShareDataDismissed(Context context) {
+        if (context == null) return false;
         SharedPreferences settingsToGet = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return settingsToGet.getBoolean("shareDataDismissed", false);
     }
@@ -351,11 +388,13 @@ public class BRSharedPrefs {
     }
 
     public static String getTrustNode(Context context) {
+        if (context == null) return "";
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         return prefs.getString("trustNode", "");
     }
 
     public static void putTrustNode(Context context, String trustNode) {
+        if (context == null) return;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putString("trustNode", trustNode);
@@ -363,35 +402,40 @@ public class BRSharedPrefs {
     }
 
     public static void incrementSendTransactionCount(Context context) {
+        if (context == null) return;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         int currentTransactions = prefs.getInt(SEND_TRANSACTION_COUNT, 0);
         prefs.edit().putInt(SEND_TRANSACTION_COUNT, currentTransactions + 1).apply();
     }
 
     public static int getSendTransactionCount(Context context) {
+        if (context == null) return 0;
         return context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE)
                 .getInt(SEND_TRANSACTION_COUNT, 0);
     }
 
     public static boolean isInAppReviewDone(Context context) {
+        if (context == null) return false;
         return context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE)
                 .getBoolean(IN_APP_REVIEW_DONE, false);
     }
 
     public static void inAppReviewDone(Context context) {
+        if (context == null) return;
         context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE)
                 .edit().putBoolean(IN_APP_REVIEW_DONE, true).apply();
     }
 
     public static float getFalsePositivesRate(Context context) {
+        if (context == null) return BRConstants.FALSE_POS_RATE_LOW_PRIVACY;
         return context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE).getFloat(PREFERRED_FPRATE, BRConstants.FALSE_POS_RATE_LOW_PRIVACY);
     }
 
     public static void putFalsePositivesRate(Context context, float preferredRate) {
+        if (context == null) return;
         SharedPreferences prefs = context.getSharedPreferences(BRConstants.PREFS_NAME, Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = prefs.edit();
         editor.putFloat(PREFERRED_FPRATE, preferredRate);
         editor.apply();
     }
 }
-
